Refactor: useTimer 추가
